requirements.txt
apache-beam==2.10.0
astroid==1.6.5
atomicwrites==1.3.0
attrs==18.2.0
avro==1.8.2
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
configparser==3.7.3
crcmod==1.7
dependency-injector==3.14.4
dill==0.2.9
docopt==0.6.2
enum34==1.1.6
fastavro==0.21.18
fasteners==0.14.1
Flask==1.0.2
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.17.1
futures==3.2.0
google==2.0.1
google-api-core==1.7.0
google-api-python-client==1.7.8
google-apitools==0.5.24
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-bigquery==1.6.1
google-cloud-core==0.29.1
google-cloud-datastore==1.7.3
google-cloud-pubsub==0.39.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
googledatastore==7.0.2
grpc-google-iam-v1==0.11.4
grpcio==1.18.0
gunicorn==19.9.0
hdfs==2.2.2
httplib2==0.9.2
idna==2.8
isort==4.3.8
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
mock==2.0.0
monotonic==1.5
more-itertools==5.0.0
numpy==1.16.2
oauth2client==3.0.0
pathlib2==2.3.3
pbr==5.1.2
pluggy==0.8.1
proto-google-cloud-datastore-v1==0.90.4
protobuf==3.6.1
py==1.7.0
pyarrow==0.11.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pydot==1.2.4
pylint==1.9.4
pyparsing==2.3.1
pyrsistent==0.14.11
pytest==4.2.1
pytest-mock==1.10.1
pytz==2018.9
PyVCF==0.6.8
PyYAML==3.13
redis==3.2.0
requests==2.21.0
rsa==4.0
scandir==1.9.0
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.8
typing==3.6.6
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.1
yml==0.0.1
